Always use node-fetch and stop polyfilling
Due to a bug in undici (https://github.com/nodejs/undici/issues/1776)
that is affecting many versions of Node and Next.js,
we want a workaround to ensure our users don't see this.

Therefore, we're forking isomorphic-unfetch to always use node-fetch
on Node to avoid falling back to global.fetch (undici).

Also, isomorphic-unfetch was polyfilling, which we want to avoid.